+++ slideshowfx-app/src/integration-test/java/com/twasyl/slideshowfx/ui/PluginCenterTest.java
@@ -1,5 +1,7 @@
 package com.twasyl.slideshowfx.ui;
 
+import com.twasyl.slideshowfx.osgi.OSGiManager;
+import com.twasyl.slideshowfx.theme.Themes;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
@@ -18,19 +20,20 @@ public class PluginCenterTest extends Application {
     @Override
     public void init() throws Exception {
         super.init();
-     //   OSGiManager.startAndDeploy();
+        OSGiManager.getInstance().startAndDeploy();
     }
 
     @Override
     public void stop() throws Exception {
         super.stop();
-       // OSGiManager.stop();
+        OSGiManager.getInstance().stop();
     }
 
     @Override
     public void start(Stage primaryStage) throws Exception {
         final Parent root = FXMLLoader.load(PluginCenterTest.class.getResource("/com/twasyl/slideshowfx/fxml/PluginCenter.fxml"));
 
+        Themes.applyTheme(root, "Dark");
         final Scene scene = new Scene(root);
         scene.getStylesheets().add("/com/twasyl/slideshowfx/css/Default.css");